Add a map() adapter.
map(e, f) produces an event e'. When e' is triggered with value v', the
system instantly triggers e with value f(v'). To get unblockers right we
need to change the signature to rendezvous::complete. You need to call map
with a signature like "map<S0>(e, f)", where S0 specifies the argument type
to f and the slot type of the returned adapter.