Native driver: Don't pointlessly call select() or gettimeofday().
I.e., call select() only if there are FDs to wait for, and call
gettimeofday() only if there are timers to wait for. Now, in most cases,
there WILL be FDs and/or timers, but not in benchmarking situations (for
measuring overhead). The pointless call to select() is in particular
expensive.