Internal: Rearrange simple_trigger code for better inlining behavior.
Before, simple_trigger() was not only large, but recursive (because of
at_trigger). Why inline all of this, when simple_trigger() contains a
call to abstract_rendezvous::complete? Having tried a handful of
arrangements, this looks the best. The linked list and at_trigger code
move into the rendezvous::completes, leaving trigger() itself nice &
small. We also make complete() auto-unuse the simple_event.