Change how active event leaks are reported.
Report the file and line of the location the event was CREATED,
not the location of the block point (which is often unset).

Simultaneously get rid of the tamer_debug_closure hack; it is really
not clear that blocking locations are meaningful on their own; if
required, one could simply look at the file and line annotations of
one or more of the associated events.